3.7.2 try... catch
I tried using a "NumberFormatException"
I had a bug where when I hit "enter," the string that was essentially a space was not read.
The thing is, when you read a value from a user, and say it's on every other line, you need to use "reader.nextline"

--> You're reading the entire line - even if it's one value that's input, and then go to the next line for the next input.

These are just nitty gritty details I need to watch out for.